Optimized reading the assets
