+++ openquake/calculators/event_based_risk.py
@@ -29,7 +29,7 @@ from openquake.baselib import (
     hdf5, performance, parallel, general, python3compat)
 from openquake.hazardlib import stats, InvalidFile
 from openquake.hazardlib.source.rupture import RuptureProxy
-from openquake.commonlib.calc import starmap_from_gmfs
+from openquake.commonlib.calc import starmap_from_gmfs, compactify3
 from openquake.risklib.scientific import (
     total_losses, insurance_losses, MultiEventRNG, LOSSID)
 from openquake.calculators import base, event_based
@@ -196,10 +196,7 @@ def event_based_risk(df, oqparam, monitor):
     :returns: a dictionary of arrays
     """
     with monitor('reading crmodel', measuremem=True):
-        tss = monitor.read('taxo-start-stop')
-        blks = [monitor.read('assets/ideductible', slice(s0, s1))
-                for t, s0, s1 in tss]
-        ideduc = numpy.concatenate(blks)
+        ideduc = monitor.read('assets/ideductible')
         aggids = monitor.read('aggids')
         crmodel = monitor.read('crmodel')
         rlz_id = monitor.read('rlz_id')
@@ -212,17 +209,16 @@ def event_based_risk(df, oqparam, monitor):
         rng = MultiEventRNG(oqparam.master_seed, df.eid.unique(),
                             int(oqparam.asset_correlation))
 
-    outs = gen_outputs(df, crmodel, tss, rng, monitor)
+    outs = gen_outputs(df, crmodel, rng, monitor)
     avg, alt = aggreg(outs, crmodel, ARK, aggids, rlz_id, ideduc.any(),
                       monitor)
     return dict(avg=avg, alt=alt)
 
 
-def gen_outputs(df, crmodel, tss, rng, monitor):
+def gen_outputs(df, crmodel, rng, monitor):
     """
     :param df: GMF dataframe (a slice of events)
     :param crmodel: CompositeRiskModel instance
-    :param tss: triples (taxonomy index, start, stop)
     :param rng: random number generator
     :param monitor: Monitor instance
     """
@@ -230,21 +226,24 @@ def gen_outputs(df, crmodel, tss, rng, monitor):
     fil_mon = monitor('filtering GMFs', measuremem=False)
     ass_mon = monitor('reading assets', measuremem=True)
     slices = performance.split_slices(df.eid.to_numpy(), 250_000)
-    for t, s0, s1 in tss:
+    for s0, s1 in monitor.read('start-stop'):
         with ass_mon:
-            adf = monitor.read('assets', slice(s0, s1)).set_index('ordinal')
-        for start, stop in slices:
-            gdf = df[start:stop]
-            with fil_mon:
-                # *crucial* for the performance of the next step
-                gmf_df = gdf[numpy.isin(gdf.sid.to_numpy(),
-                                        adf.site_id.to_numpy())]
-            if len(gmf_df) == 0:  # common enough
-                continue
-            with mon_risk:
-                out = crmodel.get_output(
-                    adf, gmf_df, crmodel.oqparam._sec_losses, rng)
-            yield out
+            assets = monitor.read('assets', slice(s0, s1)).set_index('ordinal')
+        taxos = numpy.sort(assets.taxonomy.unique())
+        for taxo in taxos:
+            adf = assets[assets.taxonomy == taxo]
+            for start, stop in slices:
+                gdf = df[start:stop]
+                with fil_mon:
+                    # *crucial* for the performance of the next step
+                    gmf_df = gdf[numpy.isin(gdf.sid.to_numpy(),
+                                            adf.site_id.to_numpy())]
+                if len(gmf_df) == 0:  # common enough
+                    continue
+                with mon_risk:
+                    out = crmodel.get_output(
+                        adf, gmf_df, crmodel.oqparam._sec_losses, rng)
+                yield out
 
 
 def ebrisk(proxies, full_lt, oqparam, dstore, monitor):
@@ -281,7 +280,7 @@ class EventBasedRiskCalculator(event_based.EventBasedCalculator):
         adf = self.assetcol.to_dframe().sort_values('taxonomy')
         monitor.save('assets', adf)
         tss = performance.idx_start_stop(adf.taxonomy.to_numpy())
-        monitor.save('taxo-start-stop', tss)
+        monitor.save('start-stop', compactify3(tss))
         monitor.save('srcfilter', srcfilter)
         monitor.save('crmodel', self.crmodel)
         monitor.save('rlz_id', self.rlzs)
